Remove mainloop stub implementations of Small Os
sol_mainloop_impl_fd_add/del and sol_mainloop_impl_child_watch_add/del
are only supported in Linux, so lets only have those functions symbols
when we have a real implementation, this way if compiling some program
that make use of that functions it will break at compile time on Small
Os not at runtime.
